+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Version 1.4.2
 - not treat every chat message starting with 'server' as a server command
 - add GOG install directory auto-detection
 - add support for late joiners aware triggers (improves Pure Faction compatibility)
+- prioritize packfile that contains rfl being loaded
 - fix sending crash reports
 - fix non-working face scrolling before first geomod use in custom levels (RF bug)
 - fix memory and VRAM leak when quick saving (RF bug)